% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars-package.R
\docType{package}
\name{mtcars-package}
\alias{mtcars}
\alias{mtcars-package}
\title{mtcars: This package does nothing}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
